refactor: extract authentication screens to shared ui package and simplify app-level registration components
