+++ apps/employer-app/assets.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../packages/koudmain-ui/src/types/assets.d.ts" />
