--- apps/employer-app/svg.d.ts
@@ -1,7 +1,0 @@
-declare module '*.svg' {
-  import React from 'react';
-  import { SvgProps } from 'react-native-svg';
-
-  const content: React.FC<SvgProps>;
-  export default content;
-}
