--- apps/worker-app/svg.d.ts
@@ -1,6 +1,0 @@
-declare module "*.svg" {
-  import React from 'react';
-  import { SvgProps } from 'react-native-svg';
-  const content: React.FC<SvgProps>;
-  export default content;
-}
