+++ packages/koudmain-ui/package.json
@@ -7,6 +7,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./gluestack": "./src/components/ui/index.ts",
+    "./screens": "./src/screens/index.ts",
     "./*": ["./src/*.tsx", "./src/*.ts"]
   },
   "scripts": {
